backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
sqlalchemy==2.0.29
alembic==1.13.1
pydantic[email]==2.6.4
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-multipart==0.0.9
google-cloud-speech==2.26.0
python-dotenv==1.0.1
websockets==12.0
aiofiles==23.2.1
httpx==0.27.0
pytest==8.3.5
celery==5.4.0
redis==5.2.1
flower==2.0.1
